requirements.txt
openai>=1.0.0
pytest>=7.0.0
keyboard>=0.13.5
python-dotenv>=1.1.1